:recycle: api: dissociating scoring algorithms and migrating it to the right context
