+++ api/src/certification/evaluation/application/jobs/certification-completed-job-controller.js
@@ -5,21 +5,13 @@ import { JobController } from '../../../../shared/application/jobs/job-controlle
 import { V3_REPRODUCIBILITY_RATE } from '../../../../shared/domain/constants.js';
 import { CertificationComputeError } from '../../../../shared/domain/errors.js';
 import { AssessmentResult } from '../../../../shared/domain/models/index.js';
-import * as answerRepository from '../../../../shared/infrastructure/repositories/answer-repository.js';
-import * as challengeRepository from '../../../../shared/infrastructure/repositories/challenge-repository.js';
-import * as flashAlgorithmService from '../../../flash-certification/domain/services/algorithm-methods/flash.js';
+import { AssessmentResultFactory } from '../../../scoring/domain/models/factories/AssessmentResultFactory.js';
 import { assessmentResultRepository } from '../../../session-management/infrastructure/repositories/index.js';
 import { AlgorithmEngineVersion } from '../../../shared/domain/models/AlgorithmEngineVersion.js';
 import * as scoringCertificationService from '../../../shared/domain/services/scoring-certification-service.js';
 import * as certificationAssessmentRepository from '../../../shared/infrastructure/repositories/certification-assessment-repository.js';
 import * as certificationCourseRepository from '../../../shared/infrastructure/repositories/certification-course-repository.js';
-import * as competenceMarkRepository from '../../../shared/infrastructure/repositories/competence-mark-repository.js';
-import * as flashAlgorithmConfigurationRepository from '../../../shared/infrastructure/repositories/flash-algorithm-configuration-repository.js';
-import { AssessmentResultFactory } from '../../domain/models/factories/AssessmentResultFactory.js';
-import { scoringDegradationService } from '../../domain/services/scoring-degradation-service.js';
-import * as certificationAssessmentHistoryRepository from '../../infrastructure/repositories/certification-assessment-history-repository.js';
-import * as certificationChallengeForScoringRepository from '../../infrastructure/repositories/certification-challenge-for-scoring-repository.js';
-import * as scoringConfigurationRepository from '../../infrastructure/repositories/scoring-configuration-repository.js';
+import { services } from '../../domain/services/index.js';
 
 export class CertificationCompletedJobController extends JobController {
   constructor() {
@@ -29,19 +21,11 @@ export class CertificationCompletedJobController extends JobController {
   async handle({
     data,
     dependencies = {
-      answerRepository,
       assessmentResultRepository,
-      certificationAssessmentHistoryRepository,
       certificationAssessmentRepository,
       certificationCourseRepository,
-      certificationChallengeForScoringRepository,
-      challengeRepository,
-      competenceMarkRepository,
-      flashAlgorithmConfigurationRepository,
-      flashAlgorithmService,
       scoringCertificationService,
-      scoringConfigurationRepository,
-      scoringDegradationService,
+      services,
       events,
     },
   }) {
@@ -49,18 +33,10 @@ export class CertificationCompletedJobController extends JobController {
 
     const {
       assessmentResultRepository,
-      certificationAssessmentHistoryRepository,
       certificationAssessmentRepository,
       certificationCourseRepository,
-      certificationChallengeForScoringRepository,
-      competenceMarkRepository,
-      scoringConfigurationRepository,
       scoringCertificationService,
-      answerRepository,
-      flashAlgorithmConfigurationRepository,
-      flashAlgorithmService,
-      scoringDegradationService,
-      challengeRepository,
+      services,
       events,
     } = dependencies;
 
@@ -71,26 +47,16 @@ export class CertificationCompletedJobController extends JobController {
       certificationScoringCompletedEvent = await _handleV3CertificationScoring({
         certificationAssessment,
         locale,
-        answerRepository,
-        assessmentResultRepository,
-        certificationAssessmentHistoryRepository,
-        certificationChallengeForScoringRepository,
         certificationCourseRepository,
-        competenceMarkRepository,
-        flashAlgorithmConfigurationRepository,
-        flashAlgorithmService,
-        scoringConfigurationRepository,
-        scoringCertificationService,
-        scoringDegradationService,
-        challengeRepository,
+        services,
       });
     } else {
       certificationScoringCompletedEvent = await _handleV2CertificationScoring({
         certificationAssessment,
         assessmentResultRepository,
         certificationCourseRepository,
-        competenceMarkRepository,
         scoringCertificationService,
+        services,
       });
     }
 
@@ -104,21 +70,16 @@ async function _handleV2CertificationScoring({
   certificationAssessment,
   assessmentResultRepository,
   certificationCourseRepository,
-  competenceMarkRepository,
   scoringCertificationService,
+  services,
 }) {
   const emitter = AssessmentResult.emitters.PIX_ALGO;
 
   try {
-    const { certificationCourse, certificationAssessmentScore } =
-      await scoringCertificationService.handleV2CertificationScoring({
-        emitter,
-        certificationAssessment,
-        assessmentResultRepository,
-        certificationCourseRepository,
-        competenceMarkRepository,
-        scoringCertificationService,
-      });
+    const { certificationCourse, certificationAssessmentScore } = await services.handleV2CertificationScoring({
+      emitter,
+      certificationAssessment,
+    });
 
     certificationCourse.complete({ now: new Date() });
 
@@ -155,35 +116,14 @@ async function _handleV2CertificationScoring({
 async function _handleV3CertificationScoring({
   certificationAssessment,
   locale,
-  answerRepository,
-  assessmentResultRepository,
-  certificationAssessmentHistoryRepository,
-  certificationChallengeForScoringRepository,
   certificationCourseRepository,
-  competenceMarkRepository,
-  flashAlgorithmConfigurationRepository,
-  flashAlgorithmService,
-  scoringConfigurationRepository,
-  scoringCertificationService,
-  scoringDegradationService,
-  challengeRepository,
+  services,
 }) {
   const emitter = AssessmentResult.emitters.PIX_ALGO;
-  const certificationCourse = await scoringCertificationService.handleV3CertificationScoring({
+  const certificationCourse = await services.handleV3CertificationScoring({
     certificationAssessment,
     emitter,
     locale,
-    answerRepository,
-    assessmentResultRepository,
-    certificationAssessmentHistoryRepository,
-    certificationChallengeForScoringRepository,
-    certificationCourseRepository,
-    competenceMarkRepository,
-    flashAlgorithmConfigurationRepository,
-    flashAlgorithmService,
-    scoringDegradationService,
-    scoringConfigurationRepository,
-    challengeRepository,
   });
 
   if (!certificationCourse.isCancelled()) {
