+++ api/src/certification/evaluation/domain/usecases/index.js
@@ -17,6 +17,7 @@ import {
   sessionRepositories,
   sharedCompetenceMarkRepository,
 } from '../../../session-management/infrastructure/repositories/index.js';
+import * as certificationCourseRepository from '../../../shared/infrastructure/repositories/certification-course-repository.js';
 import * as certificationCandidateRepository from '../../infrastructure/repositories/certification-candidate-repository.js';
 import * as certificationCompanionAlertRepository from '../../infrastructure/repositories/certification-companion-alert-repository.js';
 
@@ -39,6 +40,7 @@ const dependencies = {
   flashAlgorithmService,
   pickChallengeService,
   certificationCompanionAlertRepository,
+  certificationCourseRepository,
 };
 
 const path = dirname(fileURLToPath(import.meta.url));
