+++ api/tests/certification/shared/integration/infrastructure/repositories/scoring-configuration-repository_test.js
@@ -1,12 +1,12 @@
 import _ from 'lodash';
 
 import { knex } from '../../../../../../db/knex-database-connection.js';
-import { V3CertificationScoring } from '../../../../../../src/certification/scoring/domain/models/V3CertificationScoring.js';
+import { V3CertificationScoring } from '../../../../../../src/certification/shared/domain/models/V3CertificationScoring.js';
 import {
   getLatestByDateAndLocale,
   saveCertificationScoringConfiguration,
   saveCompetenceForScoringConfiguration,
-} from '../../../../../../src/certification/scoring/infrastructure/repositories/scoring-configuration-repository.js';
+} from '../../../../../../src/certification/shared/infrastructure/repositories/scoring-configuration-repository.js';
 import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, databaseBuilder, expect, mockLearningContent } from '../../../../../test-helper.js';
 import { buildArea, buildCompetence, buildFramework } from '../../../../../tooling/domain-builder/factory/index.js';
