+++ api/tests/certification/shared/unit/domain/services/scoring-certification-service_test.js
@@ -1,27 +1,10 @@
 import _ from 'lodash';
 
-import { CertificationCompletedJob } from '../../../../../../lib/domain/events/CertificationCompleted.js';
-import { CertificationCourseRejected } from '../../../../../../lib/domain/events/CertificationCourseRejected.js';
-import { CertificationJuryDone } from '../../../../../../lib/domain/events/CertificationJuryDone.js';
-import { ChallengeDeneutralized } from '../../../../../../lib/domain/events/ChallengeDeneutralized.js';
-import { ChallengeNeutralized } from '../../../../../../lib/domain/events/ChallengeNeutralized.js';
-import { CertificationChallengeForScoring } from '../../../../../../src/certification/scoring/domain/models/CertificationChallengeForScoring.js';
-import {
-  CertificationAssessment,
-  states,
-} from '../../../../../../src/certification/session-management/domain/models/CertificationAssessment.js';
+import { states } from '../../../../../../src/certification/session-management/domain/models/CertificationAssessment.js';
 import { ABORT_REASONS } from '../../../../../../src/certification/shared/domain/models/CertificationCourse.js';
-import { AutoJuryCommentKeys } from '../../../../../../src/certification/shared/domain/models/JuryComment.js';
-import { SESSIONS_VERSIONS } from '../../../../../../src/certification/shared/domain/models/SessionVersion.js';
 import * as scoringCertificationService from '../../../../../../src/certification/shared/domain/services/scoring-certification-service.js';
-import { config } from '../../../../../../src/shared/config.js';
 import { CertificationComputeError } from '../../../../../../src/shared/domain/errors.js';
-import { AssessmentResult, status } from '../../../../../../src/shared/domain/models/AssessmentResult.js';
-import { CertificationResult } from '../../../../../../src/shared/domain/models/index.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
-import { generateAnswersForChallenges, generateChallengeList } from '../../../fixtures/challenges.js';
-
-const { minimumAnswersRequiredToValidateACertification } = config.v3Certification.scoring;
 
 function _buildUserCompetence(competence, pixScore, estimatedLevel) {
   return domainBuilder.buildUserCompetence({ ...competence, estimatedLevel, pixScore });
@@ -230,7 +213,6 @@ const userCompetences = [
   _buildUserCompetence(competence_3, pixForCompetence3, 3),
   _buildUserCompetence(competence_4, pixForCompetence4, 4),
 ];
-const maximumAssessmentLength = 32;
 
 describe('Certification | Shared | Unit | Domain | Services | Scoring Certification Service', function () {
   context('#calculateCertificationAssessmentScore', function () {
@@ -1326,2093 +1308,4 @@ describe('Certification | Shared | Unit | Domain | Services | Scoring Certificat
       });
     });
   });
-
-  context('#handleV2CertificationScoring', function () {
-    let assessmentResultRepository,
-      competenceMarkRepository,
-      certificationCourseRepository,
-      calculateCertificationAssessmentScoreInjected;
-
-    beforeEach(function () {
-      competenceMarkRepository = { save: sinon.stub() };
-      assessmentResultRepository = { save: sinon.stub() };
-      certificationCourseRepository = { get: sinon.stub() };
-      calculateCertificationAssessmentScoreInjected = sinon.stub();
-    });
-
-    context('for scoring certification', function () {
-      context('when candidate has enough correct answers to be certified', function () {
-        let certificationCourseId, certificationAssessmentScore, certificationAssessment, certificationCourse;
-
-        beforeEach(function () {
-          const assessmentResultId = 123123;
-          certificationCourseId = 123;
-          const competenceMark1 = domainBuilder.buildCompetenceMark({ assessmentResultId, score: 5 });
-          const competenceMark2 = domainBuilder.buildCompetenceMark({ assessmentResultId, score: 4 });
-          certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScore({
-            competenceMarks: [competenceMark1, competenceMark2],
-            percentageCorrectAnswers: 80,
-          });
-          certificationAssessment = domainBuilder.buildCertificationAssessment({
-            id: 987,
-            certificationCourseId,
-            userId: 4567,
-          });
-
-          certificationCourse = domainBuilder.buildCertificationCourse({
-            id: certificationCourseId,
-            completedAt: null,
-          });
-          const savedAssessmentResult = domainBuilder.certification.scoring.buildAssessmentResult.standard({
-            id: assessmentResultId,
-            emitter: AssessmentResult.emitters.PIX_ALGO,
-            pixScore: certificationAssessmentScore.nbPix,
-            reproducibilityRate: certificationAssessmentScore.percentageCorrectAnswers,
-            status: certificationAssessmentScore.status,
-            assessmentId: certificationAssessment.id,
-          });
-
-          calculateCertificationAssessmentScoreInjected.resolves(certificationAssessmentScore);
-          certificationCourseRepository.get
-            .withArgs({ id: certificationAssessment.certificationCourseId })
-            .resolves(certificationCourse);
-          assessmentResultRepository.save
-            .withArgs({ certificationCourseId, assessmentResult: savedAssessmentResult })
-            .resolves(savedAssessmentResult);
-          competenceMarkRepository.save.resolves();
-        });
-
-        it('builds and save an assessment result with the expected arguments', async function () {
-          // when
-          await scoringCertificationService.handleV2CertificationScoring({
-            emitter: AssessmentResult.emitters.PIX_ALGO,
-            certificationAssessment,
-            assessmentResultRepository,
-            certificationCourseRepository,
-            competenceMarkRepository,
-            calculateCertificationAssessmentScoreInjected,
-          });
-
-          // then
-          const expectedAssessmentResult = new AssessmentResult({
-            pixScore: certificationAssessmentScore.nbPix,
-            reproducibilityRate: certificationAssessmentScore.percentageCorrectAnswers,
-            status: certificationAssessmentScore.status,
-            assessmentId: certificationAssessment.id,
-            emitter: AssessmentResult.emitters.PIX_ALGO,
-          });
-          expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-            certificationCourseId,
-            assessmentResult: expectedAssessmentResult,
-          });
-        });
-
-        it('builds and save as many competence marks as present in the certificationAssessmentScore', async function () {
-          // when
-          await scoringCertificationService.handleV2CertificationScoring({
-            emitter: AssessmentResult.emitters.PIX_ALGO,
-            certificationAssessment,
-            assessmentResultRepository,
-            certificationCourseRepository,
-            competenceMarkRepository,
-            calculateCertificationAssessmentScoreInjected,
-          });
-
-          // then
-          expect(competenceMarkRepository.save.callCount).to.equal(certificationAssessmentScore.competenceMarks.length);
-        });
-      });
-
-      context('when candidate has insufficient correct answers to be certified', function () {
-        it('builds and save an insufficient correct answers assessment result', async function () {
-          // given
-          const certificationCourseId = 123;
-          const certificationCourse = domainBuilder.buildCertificationCourse({
-            id: certificationCourseId,
-            abortReason: null,
-          });
-          const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScore({
-            competenceMarks: [],
-            percentageCorrectAnswers: 49,
-            hasEnoughNonNeutralizedChallengesToBeTrusted: true,
-          });
-          const certificationAssessment = domainBuilder.buildCertificationAssessment({
-            id: 45674567,
-            certificationCourseId,
-            userId: 4567,
-          });
-          const savedAssessmentResult = { id: 123123 };
-
-          calculateCertificationAssessmentScoreInjected.resolves(certificationAssessmentScore);
-          certificationCourseRepository.get
-            .withArgs({ id: certificationAssessment.certificationCourseId })
-            .resolves(certificationCourse);
-          assessmentResultRepository.save.resolves(savedAssessmentResult);
-          competenceMarkRepository.save.resolves();
-
-          // when
-          await scoringCertificationService.handleV2CertificationScoring({
-            emitter: AssessmentResult.emitters.PIX_ALGO,
-            certificationAssessment,
-            assessmentResultRepository,
-            certificationCourseRepository,
-            competenceMarkRepository,
-            calculateCertificationAssessmentScoreInjected,
-          });
-
-          // then
-          const expectedAssessmentResult = new AssessmentResult({
-            pixScore: 0,
-            reproducibilityRate: certificationAssessmentScore.getPercentageCorrectAnswers(),
-            status: status.REJECTED,
-            assessmentId: certificationAssessment.id,
-            emitter: AssessmentResult.emitters.PIX_ALGO,
-            commentForCandidate: domainBuilder.certification.shared.buildJuryComment.candidate({
-              commentByAutoJury: AutoJuryCommentKeys.REJECTED_DUE_TO_INSUFFICIENT_CORRECT_ANSWERS,
-            }),
-            commentForOrganization: domainBuilder.certification.shared.buildJuryComment.organization({
-              commentByAutoJury: AutoJuryCommentKeys.REJECTED_DUE_TO_INSUFFICIENT_CORRECT_ANSWERS,
-            }),
-          });
-
-          expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-            certificationCourseId: 123,
-            assessmentResult: expectedAssessmentResult,
-          });
-        });
-      });
-    });
-
-    context('for rescoring certification', function () {
-      it('computes and persists the assessment result and competence marks', async function () {
-        // given
-        const certificationCourse = domainBuilder.buildCertificationCourse({
-          isCancelled: false,
-        });
-
-        const event = new ChallengeNeutralized({ certificationCourseId: certificationCourse.getId(), juryId: 7 });
-        const certificationAssessment = new CertificationAssessment({
-          id: 123,
-          userId: 123,
-          certificationCourseId: certificationCourse.getId(),
-          createdAt: new Date('2020-01-01'),
-          completedAt: new Date('2020-01-01'),
-          state: CertificationAssessment.states.STARTED,
-          version: 2,
-          certificationChallenges: [
-            domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-            domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-          ],
-          certificationAnswersByDate: ['answer'],
-        });
-        const competenceMark2 = domainBuilder.buildCompetenceMark({ score: 5 });
-        const competenceMark1 = domainBuilder.buildCompetenceMark({ score: 4 });
-        const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScore({
-          nbPix: 9,
-          status: AssessmentResult.status.VALIDATED,
-          competenceMarks: [competenceMark1, competenceMark2],
-          percentageCorrectAnswers: 80,
-          hasEnoughNonNeutralizedChallengesToBeTrusted: true,
-        });
-
-        const assessmentResultToBeSaved = new AssessmentResult({
-          id: undefined,
-          emitter: 'PIX-ALGO-NEUTRALIZATION',
-          pixScore: 9,
-          reproducibilityRate: 80,
-          status: AssessmentResult.status.VALIDATED,
-          assessmentId: 123,
-          juryId: 7,
-        });
-        const savedAssessmentResult = new AssessmentResult({ ...assessmentResultToBeSaved, id: 4 });
-        assessmentResultRepository.save.resolves({
-          certificationCourseId: 789,
-          assessmentResult: savedAssessmentResult,
-        });
-
-        calculateCertificationAssessmentScoreInjected.resolves(certificationAssessmentScore);
-        certificationCourseRepository.get
-          .withArgs({ id: certificationAssessment.certificationCourseId })
-          .resolves(certificationCourse);
-        assessmentResultRepository.save.resolves(savedAssessmentResult);
-        competenceMarkRepository.save.resolves();
-
-        // when
-        await scoringCertificationService.handleV2CertificationScoring({
-          event,
-          emitter: CertificationResult.emitters.PIX_ALGO_NEUTRALIZATION,
-          certificationAssessment,
-          assessmentResultRepository,
-          certificationCourseRepository,
-          competenceMarkRepository,
-          calculateCertificationAssessmentScoreInjected,
-        });
-
-        // then
-        expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-          certificationCourseId: 123,
-          assessmentResult: assessmentResultToBeSaved,
-        });
-        competenceMark1.assessmentResultId = savedAssessmentResult.id;
-        competenceMark2.assessmentResultId = savedAssessmentResult.id;
-        expect(competenceMarkRepository.save).to.have.been.calledWithExactly(competenceMark1);
-        expect(competenceMarkRepository.save).to.have.been.calledWithExactly(competenceMark2);
-      });
-
-      context('when the certification has not enough non neutralized challenges to be trusted', function () {
-        let certificationCourse, event, certificationAssessment;
-
-        beforeEach(function () {
-          certificationCourse = domainBuilder.buildCertificationCourse({ id: 789 });
-          event = new ChallengeNeutralized({ certificationCourseId: 789, juryId: 7 });
-          certificationAssessment = new CertificationAssessment({
-            id: 123,
-            userId: 123,
-            certificationCourseId: 789,
-            createdAt: new Date('2020-01-01'),
-            completedAt: new Date('2020-01-01'),
-            state: CertificationAssessment.states.STARTED,
-            version: 2,
-            certificationChallenges: [
-              domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-              domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-            ],
-            certificationAnswersByDate: ['answer'],
-          });
-        });
-
-        it('builds and save a not trustable assessment result', async function () {
-          // given
-          const competenceMark2 = domainBuilder.buildCompetenceMark({ score: 12 });
-          const competenceMark1 = domainBuilder.buildCompetenceMark({ score: 18 });
-          const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScore({
-            status: AssessmentResult.status.VALIDATED,
-            competenceMarks: [competenceMark1, competenceMark2],
-            percentageCorrectAnswers: 80,
-            hasEnoughNonNeutralizedChallengesToBeTrusted: false,
-          });
-          const assessmentResultToBeSaved = domainBuilder.certification.scoring.buildAssessmentResult.notTrustable({
-            emitter: 'PIX-ALGO-NEUTRALIZATION',
-            pixScore: 30,
-            reproducibilityRate: 80,
-            status: AssessmentResult.status.VALIDATED,
-            assessmentId: 123,
-            juryId: 7,
-          });
-          const savedAssessmentResult = new AssessmentResult({ ...assessmentResultToBeSaved, id: 4 });
-
-          assessmentResultRepository.save.resolves({
-            certificationCourseId: 789,
-            assessmentResult: savedAssessmentResult,
-          });
-          calculateCertificationAssessmentScoreInjected.resolves(certificationAssessmentScore);
-          certificationCourseRepository.get
-            .withArgs({ id: certificationAssessment.certificationCourseId })
-            .resolves(certificationCourse);
-          assessmentResultRepository.save.resolves(savedAssessmentResult);
-          competenceMarkRepository.save.resolves();
-
-          // when
-          await scoringCertificationService.handleV2CertificationScoring({
-            event,
-            emitter: CertificationResult.emitters.PIX_ALGO_NEUTRALIZATION,
-            certificationAssessment,
-            assessmentResultRepository,
-            certificationCourseRepository,
-            competenceMarkRepository,
-            calculateCertificationAssessmentScoreInjected,
-          });
-
-          // then
-          expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-            certificationCourseId: 789,
-            assessmentResult: assessmentResultToBeSaved,
-          });
-        });
-
-        context('when it has insufficient correct answers', function () {
-          it('builds and save a not trustable assessment result', async function () {
-            // given
-            const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScore({
-              status: AssessmentResult.status.REJECTED,
-              percentageCorrectAnswers: 45,
-              hasEnoughNonNeutralizedChallengesToBeTrusted: false,
-            });
-            const assessmentResultToBeSaved = domainBuilder.certification.scoring.buildAssessmentResult.notTrustable({
-              emitter: 'PIX-ALGO-NEUTRALIZATION',
-              pixScore: 0,
-              reproducibilityRate: 45,
-              status: AssessmentResult.status.REJECTED,
-              assessmentId: 123,
-              juryId: 7,
-            });
-            const savedAssessmentResult = new AssessmentResult({ ...assessmentResultToBeSaved, id: 4 });
-
-            assessmentResultRepository.save.resolves({
-              certificationCourseId: 789,
-              assessmentResult: savedAssessmentResult,
-            });
-            calculateCertificationAssessmentScoreInjected.resolves(certificationAssessmentScore);
-            certificationCourseRepository.get
-              .withArgs({ id: certificationAssessment.certificationCourseId })
-              .resolves(certificationCourse);
-            assessmentResultRepository.save.resolves(savedAssessmentResult);
-            competenceMarkRepository.save.resolves();
-
-            // when
-            await scoringCertificationService.handleV2CertificationScoring({
-              event,
-              emitter: CertificationResult.emitters.PIX_ALGO_NEUTRALIZATION,
-              certificationAssessment,
-              assessmentResultRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              calculateCertificationAssessmentScoreInjected,
-            });
-
-            // then
-            expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-              certificationCourseId: 789,
-              assessmentResult: assessmentResultToBeSaved,
-            });
-          });
-        });
-      });
-
-      context('when the certification has enough non neutralized challenges to be trusted', function () {
-        it('build and save a standard assessment result', async function () {
-          // given
-          const certificationCourse = domainBuilder.buildCertificationCourse({ id: 789, isCancelled: true });
-
-          const event = new ChallengeNeutralized({ certificationCourseId: 789, juryId: 7 });
-          const certificationAssessment = new CertificationAssessment({
-            id: 123,
-            userId: 123,
-            certificationCourseId: 789,
-            createdAt: new Date('2020-01-01'),
-            completedAt: new Date('2020-01-01'),
-            state: CertificationAssessment.states.STARTED,
-            version: 2,
-            certificationChallenges: [
-              domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-              domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-            ],
-            certificationAnswersByDate: ['answer'],
-          });
-          certificationCourseRepository.get.withArgs({ id: 789 }).resolves(certificationCourse);
-          const competenceMark2 = domainBuilder.buildCompetenceMark({ score: 12 });
-          const competenceMark1 = domainBuilder.buildCompetenceMark({ score: 18 });
-          const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScore({
-            status: AssessmentResult.status.VALIDATED,
-            competenceMarks: [competenceMark1, competenceMark2],
-            percentageCorrectAnswers: 80,
-            hasEnoughNonNeutralizedChallengesToBeTrusted: true,
-          });
-          const assessmentResultToBeSaved = domainBuilder.certification.scoring.buildAssessmentResult.standard({
-            emitter: CertificationResult.emitters.PIX_ALGO_NEUTRALIZATION,
-            pixScore: 30,
-            reproducibilityRate: 80,
-            status: AssessmentResult.status.VALIDATED,
-            assessmentId: 123,
-            juryId: 7,
-          });
-          const savedAssessmentResult = new AssessmentResult({ ...assessmentResultToBeSaved, id: 4 });
-
-          assessmentResultRepository.save.resolves({
-            certificationCourseId: 789,
-            assessmentResult: savedAssessmentResult,
-          });
-          calculateCertificationAssessmentScoreInjected.resolves(certificationAssessmentScore);
-          certificationCourseRepository.get
-            .withArgs({ id: certificationAssessment.certificationCourseId })
-            .resolves(certificationCourse);
-          assessmentResultRepository.save.resolves(savedAssessmentResult);
-          competenceMarkRepository.save.resolves();
-
-          // when
-          await scoringCertificationService.handleV2CertificationScoring({
-            event,
-            emitter: CertificationResult.emitters.PIX_ALGO_NEUTRALIZATION,
-            certificationAssessment,
-            assessmentResultRepository,
-            certificationCourseRepository,
-            competenceMarkRepository,
-            calculateCertificationAssessmentScoreInjected,
-          });
-
-          // then
-          expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-            certificationCourseId: 789,
-            assessmentResult: assessmentResultToBeSaved,
-          });
-        });
-      });
-
-      context('when the certification course is rejected', function () {
-        context('when it is rejected for fraud', function () {
-          it('builds and save a standard rejected assessment result ', async function () {
-            // given
-            const certificationCourse = domainBuilder.buildCertificationCourse({
-              id: 789,
-              isRejectedForFraud: true,
-            });
-
-            const event = new ChallengeNeutralized({ certificationCourseId: 789, juryId: 7 });
-            const certificationAssessment = new CertificationAssessment({
-              id: 123,
-              userId: 123,
-              certificationCourseId: 789,
-              createdAt: new Date('2020-01-01'),
-              completedAt: new Date('2020-01-01'),
-              state: CertificationAssessment.states.STARTED,
-              version: 2,
-              certificationChallenges: [
-                domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-                domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-              ],
-              certificationAnswersByDate: ['answer'],
-            });
-            const competenceMark2 = domainBuilder.buildCompetenceMark({ score: 12 });
-            const competenceMark1 = domainBuilder.buildCompetenceMark({ score: 18 });
-            const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScore({
-              status: AssessmentResult.status.VALIDATED,
-              competenceMarks: [competenceMark1, competenceMark2],
-              percentageCorrectAnswers: 80,
-              hasEnoughNonNeutralizedChallengesToBeTrusted: true,
-            });
-
-            const savedAssessmentResult = domainBuilder.certification.scoring.buildAssessmentResult.fraud({
-              pixScore: 30,
-              reproducibilityRate: 80,
-              assessmentId: 123,
-              juryId: 7,
-            });
-            assessmentResultRepository.save.resolves({
-              certificationCourseId: 789,
-              assessmentResult: savedAssessmentResult,
-            });
-            calculateCertificationAssessmentScoreInjected.resolves(certificationAssessmentScore);
-            certificationCourseRepository.get
-              .withArgs({ id: certificationAssessment.certificationCourseId })
-              .resolves(certificationCourse);
-            competenceMarkRepository.save.resolves();
-
-            // when
-            await scoringCertificationService.handleV2CertificationScoring({
-              event,
-              emitter: CertificationResult.emitters.PIX_ALGO_NEUTRALIZATION,
-              certificationAssessment,
-              assessmentResultRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              calculateCertificationAssessmentScoreInjected,
-            });
-
-            // then
-            expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-              certificationCourseId: 789,
-              assessmentResult: savedAssessmentResult,
-            });
-          });
-        });
-
-        context('when it is rejected for insufficient correct answers', function () {
-          it('builds and save an insufficient correct answers assessment result', async function () {
-            // given
-            const certificationCourse = domainBuilder.buildCertificationCourse({
-              id: 789,
-            });
-
-            const event = new ChallengeDeneutralized({ certificationCourseId: 789, juryId: 7 });
-            const certificationAssessment = new CertificationAssessment({
-              id: 123,
-              userId: 123,
-              certificationCourseId: 789,
-              createdAt: new Date('2020-01-01'),
-              completedAt: new Date('2020-01-01'),
-              state: CertificationAssessment.states.ENDED_BY_SUPERVISOR,
-              version: 2,
-              certificationChallenges: [
-                domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-                domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-                domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false }),
-              ],
-              certificationAnswersByDate: ['answer'],
-            });
-            const competenceMark1 = domainBuilder.buildCompetenceMark({ score: 0 });
-            const competenceMark2 = domainBuilder.buildCompetenceMark({ score: 0 });
-            const competenceMark3 = domainBuilder.buildCompetenceMark({ score: 0 });
-            const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScore({
-              competenceMarks: [competenceMark1, competenceMark2, competenceMark3],
-              percentageCorrectAnswers: 33,
-              hasEnoughNonNeutralizedChallengesToBeTrusted: true,
-            });
-
-            const assessmentResultToBeSaved =
-              domainBuilder.certification.scoring.buildAssessmentResult.insufficientCorrectAnswers({
-                pixScore: 0,
-                reproducibilityRate: 33,
-                assessmentId: 123,
-                emitter: CertificationResult.emitters.PIX_ALGO_NEUTRALIZATION,
-                juryId: 7,
-              });
-            assessmentResultRepository.save.resolves({
-              certificationCourseId: 789,
-              assessmentResult: assessmentResultToBeSaved,
-            });
-            calculateCertificationAssessmentScoreInjected.resolves(certificationAssessmentScore);
-            certificationCourseRepository.get
-              .withArgs({ id: certificationAssessment.certificationCourseId })
-              .resolves(certificationCourse);
-            competenceMarkRepository.save.resolves();
-
-            // when
-            await scoringCertificationService.handleV2CertificationScoring({
-              event,
-              emitter: CertificationResult.emitters.PIX_ALGO_NEUTRALIZATION,
-              certificationAssessment,
-              assessmentResultRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              calculateCertificationAssessmentScoreInjected,
-            });
-
-            // then
-            expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-              certificationCourseId: 789,
-              assessmentResult: assessmentResultToBeSaved,
-            });
-          });
-
-          context('when the candidate encountered a technical issue during certification', function () {
-            it('builds and save an assessment result lacking answers for technical reason', async function () {
-              // given
-              const certificationCourse = domainBuilder.buildCertificationCourse({
-                id: 789,
-                abortReason: ABORT_REASONS.TECHNICAL,
-              });
-
-              const event = new ChallengeDeneutralized({ certificationCourseId: 789, juryId: 7 });
-              const certificationAssessment = new CertificationAssessment({
-                id: 123,
-                userId: 123,
-                certificationCourseId: 789,
-                createdAt: new Date('2020-01-01'),
-                completedAt: new Date('2020-01-01'),
-                state: CertificationAssessment.states.ENDED_BY_SUPERVISOR,
-                version: 2,
-                certificationChallenges: [domainBuilder.buildCertificationChallengeWithType({ isNeutralized: false })],
-                certificationAnswersByDate: ['answer'],
-              });
-              const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScore({
-                percentageCorrectAnswers: 33,
-                hasEnoughNonNeutralizedChallengesToBeTrusted: true,
-              });
-
-              const assessmentResultToBeSaved =
-                domainBuilder.certification.scoring.buildAssessmentResult.lackOfAnswersForTechnicalReason({
-                  pixScore: 0,
-                  reproducibilityRate: 33,
-                  status: AssessmentResult.status.REJECTED,
-                  assessmentId: 123,
-                  emitter: CertificationResult.emitters.PIX_ALGO_NEUTRALIZATION,
-                  juryId: 7,
-                });
-
-              assessmentResultRepository.save.resolves({
-                certificationCourseId: 789,
-                assessmentResult: assessmentResultToBeSaved,
-              });
-              calculateCertificationAssessmentScoreInjected.resolves(certificationAssessmentScore);
-              certificationCourseRepository.get
-                .withArgs({ id: certificationAssessment.certificationCourseId })
-                .resolves(certificationCourse);
-              competenceMarkRepository.save.resolves();
-
-              // when
-              await scoringCertificationService.handleV2CertificationScoring({
-                event,
-                emitter: CertificationResult.emitters.PIX_ALGO_NEUTRALIZATION,
-                certificationAssessment,
-                assessmentResultRepository,
-                certificationCourseRepository,
-                competenceMarkRepository,
-                calculateCertificationAssessmentScoreInjected,
-              });
-
-              // then
-              expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-                certificationCourseId: 789,
-                assessmentResult: assessmentResultToBeSaved,
-              });
-            });
-          });
-        });
-      });
-    });
-  });
-
-  context('#handleV3CertificationScoring', function () {
-    let answerRepository,
-      assessmentResultRepository,
-      certificationAssessmentHistoryRepository,
-      certificationChallengeForScoringRepository,
-      certificationCourseRepository,
-      competenceMarkRepository,
-      flashAlgorithmConfigurationRepository,
-      flashAlgorithmService,
-      scoringDegradationService,
-      scoringConfigurationRepository,
-      challengeRepository,
-      baseFlashAlgorithmConfiguration;
-    let clock;
-    const now = new Date('2019-01-01T05:06:07Z');
-
-    beforeEach(function () {
-      clock = sinon.useFakeTimers({ now, toFake: ['Date'] });
-
-      answerRepository = {
-        findByAssessment: sinon.stub(),
-      };
-      assessmentResultRepository = { save: sinon.stub() };
-      certificationAssessmentHistoryRepository = { save: sinon.stub() };
-      certificationChallengeForScoringRepository = { getByCertificationCourseId: sinon.stub() };
-      certificationCourseRepository = { get: sinon.stub() };
-      competenceMarkRepository = { save: sinon.stub() };
-      flashAlgorithmConfigurationRepository = { getMostRecentBeforeDate: sinon.stub() };
-      flashAlgorithmService = {
-        getCapacityAndErrorRate: sinon.stub(),
-        getCapacityAndErrorRateHistory: sinon.stub(),
-      };
-      scoringDegradationService = { downgradeCapacity: sinon.stub() };
-      scoringConfigurationRepository = { getLatestByDateAndLocale: sinon.stub() };
-      challengeRepository = { findFlashCompatibleWithoutLocale: sinon.stub() };
-      baseFlashAlgorithmConfiguration = domainBuilder.buildFlashAlgorithmConfiguration({
-        maximumAssessmentLength,
-      });
-    });
-
-    afterEach(function () {
-      clock.restore();
-    });
-
-    context('for scoring certification', function () {
-      let event;
-      let certificationAssessment;
-      let certificationCourse;
-      const assessmentResultId = 99;
-      const assessmentId = 1214;
-      const certificationCourseId = 1234;
-      const userId = 4567;
-      const certificationCourseStartDate = new Date('2022-02-01');
-      let scoringConfiguration;
-      let baseFlashAlgorithmConfig;
-
-      beforeEach(function () {
-        event = new CertificationCompletedJob({
-          assessmentId,
-          userId,
-          certificationCourseId: 123,
-        });
-        certificationAssessment = {
-          id: assessmentId,
-          certificationCourseId,
-          userId,
-          createdAt: Symbol('someCreationDate'),
-          version: 3,
-        };
-        certificationCourse = domainBuilder.buildCertificationCourse({
-          id: certificationCourseId,
-          createdAt: certificationCourseStartDate,
-          completedAt: null,
-        });
-
-        scoringConfiguration = domainBuilder.buildV3CertificationScoring({
-          competencesForScoring: [domainBuilder.buildCompetenceForScoring()],
-        });
-
-        scoringConfigurationRepository.getLatestByDateAndLocale
-          .withArgs({ locale: 'fr', date: certificationCourse.getStartDate() })
-          .resolves(scoringConfiguration);
-
-        assessmentResultRepository.save.resolves(
-          domainBuilder.buildAssessmentResult({
-            id: assessmentResultId,
-          }),
-        );
-        competenceMarkRepository.save.resolves();
-
-        baseFlashAlgorithmConfiguration = domainBuilder.buildFlashAlgorithmConfiguration({
-          maximumAssessmentLength,
-        });
-      });
-
-      it('should save the score', async function () {
-        // given
-        const certificationCourseStartDate = new Date('2022-01-01');
-        // TODO: switch to certif-course version, not session
-        const certificationAssessment = domainBuilder.buildCertificationAssessment({
-          version: SESSIONS_VERSIONS.V3,
-        });
-
-        const abortedCertificationCourse = domainBuilder.buildCertificationCourse({
-          createdAt: certificationCourseStartDate,
-        });
-
-        const challenges = generateChallengeList({ length: maximumAssessmentLength });
-        const certificationChallengesForScoring = challenges.map((challenge) =>
-          domainBuilder.buildCertificationChallengeForScoring(challenge),
-        );
-        const answers = generateAnswersForChallenges({ challenges });
-
-        const expectedCapacity = 2;
-        const scoreForCapacity = 438;
-        const { certificationCourseId } = certificationAssessment;
-
-        const capacityHistory = [
-          domainBuilder.buildCertificationChallengeCapacity({
-            certificationChallengeId: certificationChallengesForScoring[0].certificationChallengeId,
-            capacity: expectedCapacity,
-          }),
-        ];
-
-        const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-          capacityHistory,
-        });
-
-        certificationChallengeForScoringRepository.getByCertificationCourseId
-          .withArgs({ certificationCourseId })
-          .resolves(certificationChallengesForScoring);
-
-        scoringConfigurationRepository.getLatestByDateAndLocale
-          .withArgs({ locale: 'fr', date: abortedCertificationCourse.getStartDate() })
-          .resolves(scoringConfiguration);
-
-        flashAlgorithmConfigurationRepository.getMostRecentBeforeDate
-          .withArgs(certificationCourseStartDate)
-          .resolves(baseFlashAlgorithmConfig);
-
-        answerRepository.findByAssessment.withArgs(certificationAssessment.id).resolves(answers);
-
-        certificationCourseRepository.get
-          .withArgs({ id: certificationAssessment.certificationCourseId })
-          .resolves(abortedCertificationCourse);
-
-        flashAlgorithmService.getCapacityAndErrorRate
-          .withArgs({
-            challenges: certificationChallengesForScoring,
-            allAnswers: answers,
-            capacity: sinon.match.number,
-            variationPercent: undefined,
-            variationPercentUntil: undefined,
-            doubleMeasuresUntil: undefined,
-          })
-          .returns({
-            capacity: expectedCapacity,
-          });
-
-        flashAlgorithmService.getCapacityAndErrorRateHistory
-          .withArgs({
-            challenges: certificationChallengesForScoring,
-            allAnswers: answers,
-            capacity: sinon.match.number,
-            variationPercent: undefined,
-            variationPercentUntil: undefined,
-            doubleMeasuresUntil: undefined,
-          })
-          .returns([
-            {
-              capacity: expectedCapacity,
-            },
-          ]);
-
-        const event = new CertificationJuryDone({
-          certificationCourseId,
-        });
-
-        // when
-        await scoringCertificationService.handleV3CertificationScoring({
-          event,
-          emitter: AssessmentResult.emitters.PIX_ALGO,
-          certificationAssessment,
-          locale: 'fr',
-          answerRepository,
-          assessmentResultRepository,
-          certificationAssessmentHistoryRepository,
-          certificationChallengeForScoringRepository,
-          certificationCourseRepository,
-          competenceMarkRepository,
-          flashAlgorithmConfigurationRepository,
-          flashAlgorithmService,
-          scoringConfigurationRepository,
-          challengeRepository,
-        });
-
-        // then
-        const expectedResult = {
-          certificationCourseId,
-          assessmentResult: new AssessmentResult({
-            emitter: AssessmentResult.emitters.PIX_ALGO,
-            pixScore: scoreForCapacity,
-            reproducibilityRate: 100,
-            status: AssessmentResult.status.VALIDATED,
-            competenceMarks: [],
-            assessmentId: 123,
-          }),
-        };
-
-        expect(assessmentResultRepository.save).to.have.been.calledWith(expectedResult);
-        expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-          certificationAssessmentHistory,
-        );
-        expect(competenceMarkRepository.save).to.have.been.calledWithExactly(
-          domainBuilder.buildCompetenceMark({
-            id: undefined,
-            assessmentResultId: 99,
-            area_code: '1',
-            competenceId: 'recCompetenceId',
-            competence_code: '1.1',
-            level: 2,
-            score: 0,
-          }),
-        );
-      });
-
-      describe('when at least the minimum number of answers required by the config has been answered', function () {
-        describe('when the certification was completed', function () {
-          it('builds and save an assessment result with a validated status', async function () {
-            // given
-            const expectedCapacity = 2;
-            const scoreForCapacity = 438;
-            const challenges = _generateCertificationChallengeForScoringList({ length: maximumAssessmentLength });
-            const answers = generateAnswersForChallenges({ challenges });
-            const assessmentResultId = 123;
-
-            const capacityHistory = [
-              domainBuilder.buildCertificationChallengeCapacity({
-                certificationChallengeId: challenges[0].certificationChallengeId,
-                capacity: expectedCapacity,
-              }),
-            ];
-
-            const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-              capacityHistory,
-            });
-
-            certificationChallengeForScoringRepository.getByCertificationCourseId
-              .withArgs({ certificationCourseId })
-              .resolves(challenges);
-            answerRepository.findByAssessment.withArgs(assessmentId).resolves(answers);
-            certificationCourseRepository.get.withArgs({ id: certificationCourseId }).resolves(certificationCourse);
-            flashAlgorithmConfigurationRepository.getMostRecentBeforeDate
-              .withArgs(certificationCourseStartDate)
-              .resolves(baseFlashAlgorithmConfiguration);
-            flashAlgorithmService.getCapacityAndErrorRate
-              .withArgs({
-                challenges,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns({
-                capacity: expectedCapacity,
-              });
-            assessmentResultRepository.save.resolves(domainBuilder.buildAssessmentResult({ id: assessmentResultId }));
-
-            flashAlgorithmService.getCapacityAndErrorRateHistory
-              .withArgs({
-                challenges,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns([
-                {
-                  capacity: expectedCapacity,
-                },
-              ]);
-
-            // when
-            await scoringCertificationService.handleV3CertificationScoring({
-              event,
-              emitter: CertificationResult.emitters.PIX_ALGO,
-              certificationAssessment,
-              locale: 'fr',
-              answerRepository,
-              assessmentResultRepository,
-              certificationAssessmentHistoryRepository,
-              certificationChallengeForScoringRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              flashAlgorithmConfigurationRepository,
-              flashAlgorithmService,
-              scoringConfigurationRepository,
-              challengeRepository,
-            });
-
-            // then
-            const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScoreV3({
-              nbPix: scoreForCapacity,
-              status: status.VALIDATED,
-            });
-            const expectedAssessmentResult = new AssessmentResult({
-              pixScore: scoreForCapacity,
-              reproducibilityRate: certificationAssessmentScore.getPercentageCorrectAnswers(),
-              status: status.VALIDATED,
-              assessmentId: certificationAssessment.id,
-              emitter: AssessmentResult.emitters.PIX_ALGO,
-            });
-            expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-              certificationCourseId: 1234,
-              assessmentResult: expectedAssessmentResult,
-            });
-            expect(competenceMarkRepository.save).to.have.been.calledWithExactly(
-              domainBuilder.buildCompetenceMark({
-                id: undefined,
-                assessmentResultId: assessmentResultId,
-                area_code: '1',
-                competenceId: 'recCompetenceId',
-                competence_code: '1.1',
-                level: 2,
-                score: 0,
-              }),
-            );
-            expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-              certificationAssessmentHistory,
-            );
-          });
-
-          describe('when the certification would reach a very high score', function () {
-            it('should return the score capped based on the maximum available level when the certification was done', async function () {
-              // given
-              const expectedCapacity = 8;
-              const cappedScoreForCapacity = 895;
-              const challenges = _generateCertificationChallengeForScoringList({ length: maximumAssessmentLength });
-
-              const answers = generateAnswersForChallenges({ challenges });
-
-              const capacityHistory = [
-                domainBuilder.buildCertificationChallengeCapacity({
-                  certificationChallengeId: challenges[0].certificationChallengeId,
-                  capacity: expectedCapacity,
-                }),
-              ];
-
-              const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-                capacityHistory,
-              });
-
-              certificationChallengeForScoringRepository.getByCertificationCourseId
-                .withArgs({ certificationCourseId })
-                .resolves(challenges);
-              answerRepository.findByAssessment.withArgs(assessmentId).resolves(answers);
-              certificationCourseRepository.get.withArgs({ id: certificationCourseId }).resolves(certificationCourse);
-              flashAlgorithmConfigurationRepository.getMostRecentBeforeDate
-                .withArgs(certificationCourseStartDate)
-                .resolves(baseFlashAlgorithmConfiguration);
-              flashAlgorithmService.getCapacityAndErrorRate
-                .withArgs({
-                  challenges,
-                  allAnswers: answers,
-                  capacity: sinon.match.number,
-                  variationPercent: undefined,
-                  variationPercentUntil: undefined,
-                  doubleMeasuresUntil: undefined,
-                })
-                .returns({
-                  capacity: expectedCapacity,
-                });
-
-              flashAlgorithmService.getCapacityAndErrorRateHistory
-                .withArgs({
-                  challenges,
-                  allAnswers: answers,
-                  capacity: sinon.match.number,
-                  variationPercent: undefined,
-                  variationPercentUntil: undefined,
-                  doubleMeasuresUntil: undefined,
-                })
-                .returns([
-                  {
-                    capacity: expectedCapacity,
-                  },
-                ]);
-
-              const emitter = CertificationResult.emitters.PIX_ALGO;
-
-              // when
-              await scoringCertificationService.handleV3CertificationScoring({
-                event,
-                emitter,
-                certificationAssessment,
-                locale: 'fr',
-                answerRepository,
-                assessmentResultRepository,
-                certificationAssessmentHistoryRepository,
-                certificationChallengeForScoringRepository,
-                certificationCourseRepository,
-                competenceMarkRepository,
-                flashAlgorithmConfigurationRepository,
-                flashAlgorithmService,
-                scoringConfigurationRepository,
-                challengeRepository,
-              });
-
-              // then
-              const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScoreV3({
-                nbPix: cappedScoreForCapacity,
-                status: status.VALIDATED,
-              });
-              const expectedAssessmentResult = new AssessmentResult({
-                pixScore: cappedScoreForCapacity,
-                reproducibilityRate: certificationAssessmentScore.getPercentageCorrectAnswers(),
-                status: status.VALIDATED,
-                assessmentId: certificationAssessment.id,
-                emitter: AssessmentResult.emitters.PIX_ALGO,
-              });
-              expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-                certificationCourseId: 1234,
-                assessmentResult: expectedAssessmentResult,
-              });
-              expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-                certificationAssessmentHistory,
-              );
-            });
-          });
-        });
-      });
-    });
-
-    context('for rescoring certification', function () {
-      let baseFlashAlgorithmConfig,
-        scoringConfiguration,
-        assessmentId,
-        certificationCourseId,
-        certificationCourseStartDate,
-        event,
-        certificationAssessment;
-
-      beforeEach(function () {
-        assessmentId = 1214;
-        certificationCourseId = 1234;
-        certificationCourseStartDate = new Date('2022-01-01');
-        event = new CertificationJuryDone({
-          certificationCourseId,
-        });
-        certificationAssessment = domainBuilder.buildCertificationAssessment({
-          certificationCourseId,
-          id: assessmentId,
-          version: SESSIONS_VERSIONS.V3,
-        });
-        baseFlashAlgorithmConfig = domainBuilder.buildFlashAlgorithmConfiguration({
-          maximumAssessmentLength,
-        });
-
-        scoringConfiguration = domainBuilder.buildV3CertificationScoring({
-          competencesForScoring: [domainBuilder.buildCompetenceForScoring()],
-        });
-
-        assessmentResultRepository.save.resolves(domainBuilder.buildAssessmentResult());
-        competenceMarkRepository.save.resolves();
-      });
-
-      describe('when the minimum number of answers required by the config were NOT answered', function () {
-        describe('when the certification was not finished due to a lack of time', function () {
-          it('should save the score with a rejected status', async function () {
-            // given
-            // TODO: switch to certif-course version, not session
-            const certificationAssessment = domainBuilder.buildCertificationAssessment({
-              version: SESSIONS_VERSIONS.V3,
-            });
-
-            const abortedCertificationCourse = domainBuilder.buildCertificationCourse({
-              abortReason: ABORT_REASONS.CANDIDATE,
-            });
-
-            const challenges = generateChallengeList({ length: minimumAnswersRequiredToValidateACertification - 1 });
-            const certificationChallengesForScoring = challenges.map((challenge) =>
-              domainBuilder.buildCertificationChallengeForScoring(challenge),
-            );
-            const answers = generateAnswersForChallenges({ challenges });
-
-            const expectedCapacity = 2;
-            const scoreForCapacity = 438;
-            const { certificationCourseId } = certificationAssessment;
-
-            const capacityHistory = [
-              domainBuilder.buildCertificationChallengeCapacity({
-                certificationChallengeId: certificationChallengesForScoring[0].certificationChallengeId,
-                capacity: expectedCapacity,
-              }),
-            ];
-
-            const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-              capacityHistory,
-            });
-
-            certificationChallengeForScoringRepository.getByCertificationCourseId
-              .withArgs({ certificationCourseId })
-              .resolves(certificationChallengesForScoring);
-
-            answerRepository.findByAssessment.withArgs(certificationAssessment.id).resolves(answers);
-
-            certificationCourseRepository.get
-              .withArgs({ id: certificationAssessment.certificationCourseId })
-              .resolves(abortedCertificationCourse);
-
-            scoringConfigurationRepository.getLatestByDateAndLocale
-              .withArgs({ locale: 'fr', date: abortedCertificationCourse.getStartDate() })
-              .resolves(scoringConfiguration);
-
-            flashAlgorithmService.getCapacityAndErrorRate
-              .withArgs({
-                challenges: certificationChallengesForScoring,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns({
-                capacity: expectedCapacity,
-              });
-
-            flashAlgorithmService.getCapacityAndErrorRateHistory
-              .withArgs({
-                challenges: certificationChallengesForScoring,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns([
-                {
-                  capacity: expectedCapacity,
-                },
-              ]);
-
-            const event = new CertificationJuryDone({
-              certificationCourseId,
-            });
-            const emitter = CertificationResult.emitters.PIX_ALGO_AUTO_JURY;
-
-            // when
-            await scoringCertificationService.handleV3CertificationScoring({
-              event,
-              emitter,
-              certificationAssessment,
-              locale: 'fr',
-              answerRepository,
-              assessmentResultRepository,
-              certificationAssessmentHistoryRepository,
-              certificationChallengeForScoringRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              flashAlgorithmConfigurationRepository,
-              flashAlgorithmService,
-              scoringConfigurationRepository,
-              challengeRepository,
-            });
-
-            // then
-            const expectedResult = {
-              certificationCourseId,
-              assessmentResult: new AssessmentResult({
-                emitter: CertificationResult.emitters.PIX_ALGO_AUTO_JURY,
-                pixScore: scoreForCapacity,
-                reproducibilityRate: 100,
-                status: AssessmentResult.status.REJECTED,
-                competenceMarks: [],
-                assessmentId: 123,
-                commentForCandidate: domainBuilder.certification.shared.buildJuryComment.candidate({
-                  commentByAutoJury: AutoJuryCommentKeys.REJECTED_DUE_TO_LACK_OF_ANSWERS,
-                }),
-                commentForOrganization: domainBuilder.certification.shared.buildJuryComment.organization({
-                  commentByAutoJury: AutoJuryCommentKeys.REJECTED_DUE_TO_LACK_OF_ANSWERS,
-                }),
-              }),
-            };
-
-            expect(assessmentResultRepository.save).to.have.been.calledWith(expectedResult);
-            expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-              certificationAssessmentHistory,
-            );
-          });
-
-          it('builds and save a lack of answers assessment result', async function () {
-            // given
-            const expectedCapacity = 2;
-            const scoreForCapacity = 438;
-            const abortedCertificationCourse = domainBuilder.buildCertificationCourse({
-              id: certificationCourseId,
-              createdAt: certificationCourseStartDate,
-              abortReason: ABORT_REASONS.CANDIDATE,
-            });
-
-            const emitter = CertificationResult.emitters.PIX_ALGO_AUTO_JURY;
-
-            const challenges = _generateCertificationChallengeForScoringList({
-              length: minimumAnswersRequiredToValidateACertification - 1,
-            });
-
-            const capacityHistory = [
-              domainBuilder.buildCertificationChallengeCapacity({
-                certificationChallengeId: challenges[0].certificationChallengeId,
-                capacity: expectedCapacity,
-              }),
-            ];
-
-            const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-              capacityHistory,
-            });
-
-            const answers = generateAnswersForChallenges({ challenges });
-
-            flashAlgorithmConfigurationRepository.getMostRecentBeforeDate
-              .withArgs(certificationCourseStartDate)
-              .resolves(baseFlashAlgorithmConfiguration);
-
-            certificationChallengeForScoringRepository.getByCertificationCourseId
-              .withArgs({ certificationCourseId })
-              .resolves(challenges);
-            answerRepository.findByAssessment.withArgs(assessmentId).resolves(answers);
-            certificationCourseRepository.get
-              .withArgs({ id: certificationCourseId })
-              .resolves(abortedCertificationCourse);
-
-            flashAlgorithmService.getCapacityAndErrorRate
-              .withArgs({
-                challenges,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns({
-                capacity: expectedCapacity,
-              });
-
-            flashAlgorithmService.getCapacityAndErrorRateHistory
-              .withArgs({
-                challenges,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns([
-                {
-                  capacity: expectedCapacity,
-                },
-              ]);
-            scoringConfigurationRepository.getLatestByDateAndLocale
-              .withArgs({ locale: 'fr', date: abortedCertificationCourse.getStartDate() })
-              .resolves(scoringConfiguration);
-
-            // when
-            await scoringCertificationService.handleV3CertificationScoring({
-              event,
-              emitter,
-              certificationAssessment,
-              locale: 'fr',
-              answerRepository,
-              assessmentResultRepository,
-              certificationAssessmentHistoryRepository,
-              certificationChallengeForScoringRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              flashAlgorithmConfigurationRepository,
-              flashAlgorithmService,
-              scoringConfigurationRepository,
-              challengeRepository,
-            });
-
-            // then
-            const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScoreV3({
-              nbPix: scoreForCapacity,
-              status: status.REJECTED,
-            });
-            const expectedAssessmentResult = new AssessmentResult({
-              pixScore: scoreForCapacity,
-              reproducibilityRate: certificationAssessmentScore.getPercentageCorrectAnswers(),
-              status: status.REJECTED,
-              assessmentId: certificationAssessment.id,
-              emitter: CertificationResult.emitters.PIX_ALGO_AUTO_JURY,
-              commentForCandidate: domainBuilder.certification.shared.buildJuryComment.candidate({
-                commentByAutoJury: AutoJuryCommentKeys.REJECTED_DUE_TO_LACK_OF_ANSWERS,
-              }),
-              commentForOrganization: domainBuilder.certification.shared.buildJuryComment.organization({
-                commentByAutoJury: AutoJuryCommentKeys.REJECTED_DUE_TO_LACK_OF_ANSWERS,
-              }),
-            });
-
-            expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-              certificationCourseId: 1234,
-              assessmentResult: expectedAssessmentResult,
-            });
-            expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-              certificationAssessmentHistory,
-            );
-          });
-        });
-
-        describe('when the certification was not finished due to technical difficulties', function () {
-          it('should save the score with a rejected status and cancel the certification course', async function () {
-            // given
-            // TODO: switch to certif-course version, not session
-            const certificationAssessment = domainBuilder.buildCertificationAssessment({
-              version: SESSIONS_VERSIONS.V3,
-            });
-
-            const abortedCertificationCourse = domainBuilder.buildCertificationCourse({
-              abortReason: ABORT_REASONS.TECHNICAL,
-            });
-
-            const challenges = generateChallengeList({ length: minimumAnswersRequiredToValidateACertification - 1 });
-            const certificationChallengesForScoring = challenges.map((challenge) =>
-              domainBuilder.buildCertificationChallengeForScoring(challenge),
-            );
-            const answers = generateAnswersForChallenges({ challenges });
-
-            const expectedCapacity = 2;
-            const scoreForCapacity = 438;
-            const { certificationCourseId } = certificationAssessment;
-
-            const capacityHistory = [
-              domainBuilder.buildCertificationChallengeCapacity({
-                certificationChallengeId: certificationChallengesForScoring[0].certificationChallengeId,
-                capacity: expectedCapacity,
-              }),
-            ];
-
-            const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-              capacityHistory,
-            });
-
-            certificationChallengeForScoringRepository.getByCertificationCourseId
-              .withArgs({ certificationCourseId })
-              .resolves(certificationChallengesForScoring);
-
-            scoringConfigurationRepository.getLatestByDateAndLocale
-              .withArgs({ locale: 'fr', date: abortedCertificationCourse.getStartDate() })
-              .resolves(scoringConfiguration);
-
-            answerRepository.findByAssessment.withArgs(certificationAssessment.id).resolves(answers);
-
-            certificationCourseRepository.get
-              .withArgs({ id: certificationAssessment.certificationCourseId })
-              .resolves(abortedCertificationCourse);
-
-            flashAlgorithmService.getCapacityAndErrorRate
-              .withArgs({
-                challenges: certificationChallengesForScoring,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns({
-                capacity: expectedCapacity,
-              });
-
-            flashAlgorithmService.getCapacityAndErrorRateHistory
-              .withArgs({
-                challenges: certificationChallengesForScoring,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns([
-                {
-                  capacity: expectedCapacity,
-                },
-              ]);
-
-            const event = new CertificationJuryDone({
-              certificationCourseId,
-            });
-            const emitter = CertificationResult.emitters.PIX_ALGO_AUTO_JURY;
-
-            // when
-            await scoringCertificationService.handleV3CertificationScoring({
-              event,
-              emitter,
-              certificationAssessment,
-              locale: 'fr',
-              answerRepository,
-              assessmentResultRepository,
-              certificationAssessmentHistoryRepository,
-              certificationChallengeForScoringRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              flashAlgorithmConfigurationRepository,
-              flashAlgorithmService,
-              scoringConfigurationRepository,
-              challengeRepository,
-            });
-
-            // then
-            const expectedResult = {
-              certificationCourseId,
-              assessmentResult: new AssessmentResult({
-                emitter: CertificationResult.emitters.PIX_ALGO_AUTO_JURY,
-                pixScore: scoreForCapacity,
-                reproducibilityRate: 100,
-                status: AssessmentResult.status.REJECTED,
-                competenceMarks: [],
-                assessmentId: 123,
-                commentForCandidate: domainBuilder.certification.shared.buildJuryComment.candidate({
-                  commentByAutoJury: AutoJuryCommentKeys.CANCELLED_DUE_TO_LACK_OF_ANSWERS_FOR_TECHNICAL_REASON,
-                }),
-                commentForOrganization: domainBuilder.certification.shared.buildJuryComment.organization({
-                  commentByAutoJury: AutoJuryCommentKeys.CANCELLED_DUE_TO_LACK_OF_ANSWERS_FOR_TECHNICAL_REASON,
-                }),
-              }),
-            };
-            expect(assessmentResultRepository.save).to.have.been.calledWith(expectedResult);
-            expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-              certificationAssessmentHistory,
-            );
-          });
-
-          it('should reject the assessment result', async function () {
-            // given
-            const abortReason = ABORT_REASONS.TECHNICAL;
-            const emitter = CertificationResult.emitters.PIX_ALGO_AUTO_JURY;
-            const expectedCapacity = 2;
-            const scoreForCapacity = 438;
-            const abortedCertificationCourse = domainBuilder.buildCertificationCourse({
-              id: certificationCourseId,
-              createdAt: certificationCourseStartDate,
-              completedAt: null,
-              abortReason,
-            });
-
-            const challenges = _generateCertificationChallengeForScoringList({
-              length: minimumAnswersRequiredToValidateACertification - 1,
-            });
-
-            const capacityHistory = [
-              domainBuilder.buildCertificationChallengeCapacity({
-                certificationChallengeId: challenges[0].certificationChallengeId,
-                capacity: expectedCapacity,
-              }),
-            ];
-
-            const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-              capacityHistory,
-            });
-
-            const answers = generateAnswersForChallenges({ challenges });
-
-            flashAlgorithmConfigurationRepository.getMostRecentBeforeDate
-              .withArgs(certificationCourseStartDate)
-              .resolves(baseFlashAlgorithmConfiguration);
-
-            certificationChallengeForScoringRepository.getByCertificationCourseId
-              .withArgs({ certificationCourseId })
-              .resolves(challenges);
-
-            answerRepository.findByAssessment.withArgs(assessmentId).resolves(answers);
-            certificationCourseRepository.get
-              .withArgs({ id: certificationCourseId })
-              .resolves(abortedCertificationCourse);
-
-            scoringConfigurationRepository.getLatestByDateAndLocale
-              .withArgs({ locale: 'fr', date: abortedCertificationCourse.getStartDate() })
-              .resolves(scoringConfiguration);
-
-            flashAlgorithmService.getCapacityAndErrorRate
-              .withArgs({
-                challenges,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns({
-                capacity: expectedCapacity,
-              });
-
-            flashAlgorithmService.getCapacityAndErrorRateHistory
-              .withArgs({
-                challenges,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns([
-                {
-                  capacity: expectedCapacity,
-                },
-              ]);
-
-            // when
-            await scoringCertificationService.handleV3CertificationScoring({
-              event,
-              emitter,
-              certificationAssessment,
-              locale: 'fr',
-              answerRepository,
-              assessmentResultRepository,
-              certificationAssessmentHistoryRepository,
-              certificationChallengeForScoringRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              flashAlgorithmConfigurationRepository,
-              flashAlgorithmService,
-              scoringConfigurationRepository,
-              challengeRepository,
-            });
-
-            // then
-            const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScoreV3({
-              nbPix: scoreForCapacity,
-              status: status.REJECTED,
-            });
-            const expectedAssessmentResult = new AssessmentResult({
-              pixScore: scoreForCapacity,
-              reproducibilityRate: certificationAssessmentScore.getPercentageCorrectAnswers(),
-              status: status.REJECTED,
-              assessmentId: certificationAssessment.id,
-              emitter: CertificationResult.emitters.PIX_ALGO_AUTO_JURY,
-              commentForCandidate: domainBuilder.certification.shared.buildJuryComment.candidate({
-                commentByAutoJury: AutoJuryCommentKeys.CANCELLED_DUE_TO_LACK_OF_ANSWERS_FOR_TECHNICAL_REASON,
-              }),
-              commentForOrganization: domainBuilder.certification.shared.buildJuryComment.organization({
-                commentByAutoJury: AutoJuryCommentKeys.CANCELLED_DUE_TO_LACK_OF_ANSWERS_FOR_TECHNICAL_REASON,
-              }),
-            });
-
-            expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-              certificationCourseId: 1234,
-              assessmentResult: expectedAssessmentResult,
-            });
-
-            expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-              certificationAssessmentHistory,
-            );
-          });
-        });
-      });
-
-      describe('when the minimum number of answers required by the config were answered', function () {
-        describe('when the candidate did not finish in time', function () {
-          it('should build and save an assessment result with a validated status', async function () {
-            // given
-            const expectedCapacity = 2;
-            const pixScore = 438;
-            const challenges = _generateCertificationChallengeForScoringList({
-              length: minimumAnswersRequiredToValidateACertification,
-            });
-            const abortReason = ABORT_REASONS.CANDIDATE;
-            const abortedCertificationCourse = domainBuilder.buildCertificationCourse({
-              id: certificationCourseId,
-              createdAt: certificationCourseStartDate,
-              completedAt: null,
-              abortReason,
-            });
-            const emitter = CertificationResult.emitters.PIX_ALGO_AUTO_JURY;
-
-            const answers = generateAnswersForChallenges({ challenges });
-
-            const capacityHistory = [
-              domainBuilder.buildCertificationChallengeCapacity({
-                certificationChallengeId: challenges[0].certificationChallengeId,
-                capacity: expectedCapacity,
-              }),
-            ];
-
-            const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-              capacityHistory,
-            });
-
-            certificationChallengeForScoringRepository.getByCertificationCourseId
-              .withArgs({ certificationCourseId })
-              .resolves(challenges);
-            answerRepository.findByAssessment.withArgs(assessmentId).resolves(answers);
-            certificationCourseRepository.get
-              .withArgs({ id: certificationCourseId })
-              .resolves(abortedCertificationCourse);
-            flashAlgorithmConfigurationRepository.getMostRecentBeforeDate
-              .withArgs(certificationCourseStartDate)
-              .resolves(baseFlashAlgorithmConfiguration);
-            flashAlgorithmService.getCapacityAndErrorRate
-              .withArgs({
-                challenges,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns({
-                capacity: expectedCapacity,
-              });
-            flashAlgorithmService.getCapacityAndErrorRateHistory
-              .withArgs({
-                challenges,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns([
-                {
-                  capacity: expectedCapacity,
-                },
-              ]);
-            challengeRepository.findFlashCompatibleWithoutLocale
-              .withArgs({
-                useObsoleteChallenges: true,
-              })
-              .returns(challenges);
-            scoringConfigurationRepository.getLatestByDateAndLocale
-              .withArgs({ locale: 'fr', date: abortedCertificationCourse.getStartDate() })
-              .resolves(scoringConfiguration);
-
-            scoringDegradationService.downgradeCapacity.returns(expectedCapacity);
-
-            // when
-            await scoringCertificationService.handleV3CertificationScoring({
-              event,
-              emitter,
-              certificationAssessment,
-              locale: 'fr',
-              answerRepository,
-              assessmentResultRepository,
-              certificationAssessmentHistoryRepository,
-              certificationChallengeForScoringRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              flashAlgorithmConfigurationRepository,
-              flashAlgorithmService,
-              scoringConfigurationRepository,
-              challengeRepository,
-              scoringDegradationService,
-            });
-
-            // then
-            const certificationAssessmentScore = domainBuilder.buildCertificationAssessmentScoreV3({
-              nbPix: pixScore,
-              status: status.VALIDATED,
-            });
-            const expectedAssessmentResult = new AssessmentResult({
-              pixScore,
-              reproducibilityRate: certificationAssessmentScore.getPercentageCorrectAnswers(),
-              status: status.VALIDATED,
-              assessmentId: certificationAssessment.id,
-              emitter: CertificationResult.emitters.PIX_ALGO_AUTO_JURY,
-            });
-            expect(assessmentResultRepository.save).to.have.been.calledWithExactly({
-              certificationCourseId,
-              assessmentResult: expectedAssessmentResult,
-            });
-            expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-              certificationAssessmentHistory,
-            );
-          });
-        });
-
-        describe('when the candidate did not finish due to technical difficulties', function () {
-          it('should save the raw score', async function () {
-            // given
-            const certificationCourseStartDate = new Date('2022-01-01');
-            // TODO: switch to certif-course version, not session
-            const certificationAssessment = domainBuilder.buildCertificationAssessment({
-              version: SESSIONS_VERSIONS.V3,
-            });
-
-            const abortedCertificationCourse = domainBuilder.buildCertificationCourse({
-              abortReason: ABORT_REASONS.TECHNICAL,
-              createdAt: certificationCourseStartDate,
-            });
-
-            const challenges = generateChallengeList({ length: minimumAnswersRequiredToValidateACertification });
-            const certificationChallengesForScoring = challenges.map((challenge) =>
-              domainBuilder.buildCertificationChallengeForScoring(challenge),
-            );
-            const answers = generateAnswersForChallenges({ challenges });
-
-            const expectedCapacity = 2;
-            const rawScore = 438;
-            const { certificationCourseId } = certificationAssessment;
-
-            const capacityHistory = [
-              domainBuilder.buildCertificationChallengeCapacity({
-                certificationChallengeId: certificationChallengesForScoring[0].certificationChallengeId,
-                capacity: expectedCapacity,
-              }),
-            ];
-
-            const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-              capacityHistory,
-            });
-
-            certificationChallengeForScoringRepository.getByCertificationCourseId
-              .withArgs({ certificationCourseId })
-              .resolves(certificationChallengesForScoring);
-
-            scoringConfigurationRepository.getLatestByDateAndLocale
-              .withArgs({ locale: 'fr', date: abortedCertificationCourse.getStartDate() })
-              .resolves(scoringConfiguration);
-
-            answerRepository.findByAssessment.withArgs(certificationAssessment.id).resolves(answers);
-
-            certificationCourseRepository.get
-              .withArgs({ id: certificationAssessment.certificationCourseId })
-              .resolves(abortedCertificationCourse);
-
-            flashAlgorithmConfigurationRepository.getMostRecentBeforeDate
-              .withArgs(abortedCertificationCourse.getStartDate())
-              .resolves(baseFlashAlgorithmConfig);
-
-            flashAlgorithmService.getCapacityAndErrorRate
-              .withArgs({
-                challenges: certificationChallengesForScoring,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns({
-                capacity: expectedCapacity,
-              });
-
-            flashAlgorithmService.getCapacityAndErrorRateHistory
-              .withArgs({
-                challenges: certificationChallengesForScoring,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns([
-                {
-                  capacity: expectedCapacity,
-                },
-              ]);
-
-            const event = new CertificationJuryDone({
-              certificationCourseId,
-            });
-            const emitter = CertificationResult.emitters.PIX_ALGO_AUTO_JURY;
-
-            // when
-            await scoringCertificationService.handleV3CertificationScoring({
-              event,
-              emitter,
-              certificationAssessment,
-              locale: 'fr',
-              answerRepository,
-              assessmentResultRepository,
-              certificationAssessmentHistoryRepository,
-              certificationChallengeForScoringRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              flashAlgorithmConfigurationRepository,
-              flashAlgorithmService,
-              scoringConfigurationRepository,
-              challengeRepository,
-            });
-
-            // then
-            const expectedResult = {
-              certificationCourseId,
-              assessmentResult: new AssessmentResult({
-                emitter: CertificationResult.emitters.PIX_ALGO_AUTO_JURY,
-                pixScore: rawScore,
-                reproducibilityRate: 100,
-                status: AssessmentResult.status.VALIDATED,
-                competenceMarks: [],
-                assessmentId: 123,
-              }),
-            };
-
-            expect(assessmentResultRepository.save).to.have.been.calledWith(expectedResult);
-            expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-              certificationAssessmentHistory,
-            );
-          });
-        });
-      });
-
-      describe('when all questions were answered', function () {
-        describe('when certification is rejected for fraud', function () {
-          it('should save the score with rejected status', async function () {
-            const certificationCourseStartDate = new Date('2022-01-01');
-            // given
-            // TODO: switch to certif-course version, not session
-            const certificationAssessment = domainBuilder.buildCertificationAssessment({
-              version: SESSIONS_VERSIONS.V3,
-            });
-
-            const abortedCertificationCourse = domainBuilder.buildCertificationCourse({
-              isRejectedForFraud: true,
-              createdAt: certificationCourseStartDate,
-            });
-
-            const challenges = generateChallengeList({ length: maximumAssessmentLength });
-            const certificationChallengesForScoring = challenges.map((challenge) =>
-              domainBuilder.buildCertificationChallengeForScoring(challenge),
-            );
-            const answers = generateAnswersForChallenges({ challenges });
-
-            const expectedCapacity = 2;
-            const scoreForCapacity = 438;
-            const { certificationCourseId } = certificationAssessment;
-
-            const capacityHistory = [
-              domainBuilder.buildCertificationChallengeCapacity({
-                certificationChallengeId: certificationChallengesForScoring[0].certificationChallengeId,
-                capacity: expectedCapacity,
-              }),
-            ];
-
-            const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-              capacityHistory,
-            });
-
-            certificationChallengeForScoringRepository.getByCertificationCourseId
-              .withArgs({ certificationCourseId })
-              .resolves(certificationChallengesForScoring);
-
-            scoringConfigurationRepository.getLatestByDateAndLocale
-              .withArgs({ locale: 'fr', date: abortedCertificationCourse.getStartDate() })
-              .resolves(scoringConfiguration);
-
-            flashAlgorithmConfigurationRepository.getMostRecentBeforeDate
-              .withArgs(certificationCourseStartDate)
-              .resolves(baseFlashAlgorithmConfig);
-
-            answerRepository.findByAssessment.withArgs(certificationAssessment.id).resolves(answers);
-
-            certificationCourseRepository.get
-              .withArgs({ id: certificationAssessment.certificationCourseId })
-              .resolves(abortedCertificationCourse);
-
-            flashAlgorithmService.getCapacityAndErrorRate
-              .withArgs({
-                challenges: certificationChallengesForScoring,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns({
-                capacity: expectedCapacity,
-              });
-
-            flashAlgorithmService.getCapacityAndErrorRateHistory
-              .withArgs({
-                challenges: certificationChallengesForScoring,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns([
-                {
-                  capacity: expectedCapacity,
-                },
-              ]);
-
-            const event = new CertificationCourseRejected({
-              certificationCourseId,
-              juryId: 7,
-            });
-
-            // when
-            await scoringCertificationService.handleV3CertificationScoring({
-              event,
-              certificationAssessment,
-              locale: 'fr',
-              answerRepository,
-              assessmentResultRepository,
-              certificationAssessmentHistoryRepository,
-              certificationChallengeForScoringRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              flashAlgorithmConfigurationRepository,
-              flashAlgorithmService,
-              scoringConfigurationRepository,
-              challengeRepository,
-            });
-
-            // then
-            const assessmentResultToBeSaved = domainBuilder.certification.scoring.buildAssessmentResult.fraud({
-              pixScore: scoreForCapacity,
-              reproducibilityRate: 100,
-              assessmentId: 123,
-              juryId: 7,
-            });
-
-            expect(assessmentResultRepository.save).to.have.been.calledWith({
-              certificationCourseId: 123,
-              assessmentResult: assessmentResultToBeSaved,
-            });
-            expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-              certificationAssessmentHistory,
-            );
-          });
-        });
-
-        describe('when the certification would reach a very high score', function () {
-          it('should return the score capped based on the maximum available level when the certification was done', async function () {
-            // given
-            const certificationCourseStartDate = new Date('2022-01-01');
-
-            // TODO: switch to certif-course version, not session
-            const certificationAssessment = domainBuilder.buildCertificationAssessment({
-              version: SESSIONS_VERSIONS.V3,
-            });
-
-            const abortedCertificationCourse = domainBuilder.buildCertificationCourse({
-              createdAt: certificationCourseStartDate,
-            });
-
-            const challenges = generateChallengeList({ length: maximumAssessmentLength });
-            const certificationChallengesForScoring = challenges.map((challenge) =>
-              domainBuilder.buildCertificationChallengeForScoring(challenge),
-            );
-
-            const answers = generateAnswersForChallenges({ challenges });
-
-            const expectedCapacity = 8;
-            const cappedscoreForCapacity = 895;
-            const { certificationCourseId } = certificationAssessment;
-
-            const capacityHistory = [
-              domainBuilder.buildCertificationChallengeCapacity({
-                certificationChallengeId: certificationChallengesForScoring[0].certificationChallengeId,
-                capacity: expectedCapacity,
-              }),
-            ];
-
-            const certificationAssessmentHistory = domainBuilder.buildCertificationAssessmentHistory({
-              capacityHistory,
-            });
-
-            certificationChallengeForScoringRepository.getByCertificationCourseId
-              .withArgs({ certificationCourseId })
-              .resolves(certificationChallengesForScoring);
-
-            scoringConfigurationRepository.getLatestByDateAndLocale
-              .withArgs({ locale: 'fr', date: abortedCertificationCourse.getStartDate() })
-              .resolves(scoringConfiguration);
-
-            flashAlgorithmConfigurationRepository.getMostRecentBeforeDate
-              .withArgs(certificationCourseStartDate)
-              .resolves(baseFlashAlgorithmConfig);
-
-            answerRepository.findByAssessment.withArgs(certificationAssessment.id).resolves(answers);
-
-            certificationCourseRepository.get
-              .withArgs({ id: certificationAssessment.certificationCourseId })
-              .resolves(abortedCertificationCourse);
-
-            flashAlgorithmService.getCapacityAndErrorRate
-              .withArgs({
-                challenges: certificationChallengesForScoring,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns({
-                capacity: expectedCapacity,
-              });
-
-            flashAlgorithmService.getCapacityAndErrorRateHistory
-              .withArgs({
-                challenges: certificationChallengesForScoring,
-                allAnswers: answers,
-                capacity: sinon.match.number,
-                variationPercent: undefined,
-                variationPercentUntil: undefined,
-                doubleMeasuresUntil: undefined,
-              })
-              .returns([
-                {
-                  capacity: expectedCapacity,
-                },
-              ]);
-
-            const expectedResult = {
-              certificationCourseId,
-              assessmentResult: new AssessmentResult({
-                emitter: CertificationResult.emitters.PIX_ALGO_AUTO_JURY,
-                pixScore: cappedscoreForCapacity,
-                reproducibilityRate: 100,
-                status: AssessmentResult.status.VALIDATED,
-                competenceMarks: [],
-                assessmentId: 123,
-              }),
-            };
-
-            const event = new CertificationJuryDone({
-              certificationCourseId,
-            });
-            const emitter = CertificationResult.emitters.PIX_ALGO_AUTO_JURY;
-
-            // when
-            await scoringCertificationService.handleV3CertificationScoring({
-              event,
-              emitter,
-              certificationAssessment,
-              locale: 'fr',
-              answerRepository,
-              assessmentResultRepository,
-              certificationAssessmentHistoryRepository,
-              certificationChallengeForScoringRepository,
-              certificationCourseRepository,
-              competenceMarkRepository,
-              flashAlgorithmConfigurationRepository,
-              flashAlgorithmService,
-              scoringConfigurationRepository,
-              challengeRepository,
-            });
-
-            // then
-            expect(assessmentResultRepository.save).to.have.been.calledWith(expectedResult);
-            expect(certificationAssessmentHistoryRepository.save).to.have.been.calledWithExactly(
-              certificationAssessmentHistory,
-            );
-          });
-        });
-      });
-    });
-  });
 });
-
-const _generateCertificationChallengeForScoringList = ({ length }) => {
-  return generateChallengeList({
-    length,
-  }).map(
-    ({ discriminant, difficulty }, index) =>
-      new CertificationChallengeForScoring({
-        certificationChallengeId: `certification-challenge-id-${index}`,
-        discriminant,
-        difficulty,
-      }),
-  );
-};
